+++ packages/egs/src/elements/attributes/BufferAttribute.ts
@@ -23,7 +23,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * The geometry attribute data store as TypedArray.
      */
-    public _array: T; // todo we should move this array's ownership into content api
+    _array: T; // todo we should move this array's ownership into content api
     /**
      * The geometry attribute data store as TypedArray.
      */
@@ -48,22 +48,22 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * If the buffer is storing a 3-component vector (such as a position, normal, or color),
      * then this will count the number of such vectors stored.
      */
-    public itemSize: number;
+    itemSize: number;
     /**
      * If attribute data is normalized.
      */
-    public normalized: boolean = false;
+    normalized: boolean = false;
     /**
      * Stores the {@link array | array's } length divided by the {@link itemSize | itemSize }.<br /><br />
      * If the buffer is storing a 3-component vector (such as a position, normal, or color),
      * then this will count the number of such vectors stored.
      */
-    public count: number;
+    count: number;
     /**
      * The data of this buffer always need update.
      * @deprecated will not take effect, update `attribute.array` if need to reupload
      */
-    public dynamic = false;
+    dynamic = false;
     /**
      * Object containing:<br />
      * offset:Integer - Default is *0*. Position at which to start update.<br />
@@ -71,26 +71,26 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * This can be used to only update some components of stored vectors (for example, just the component related to color).
      * @deprecated will not take effect, update `attribute.array` instead
      */
-    public updateRange = { offset: 0, count: - 1 };
+    updateRange = { offset: 0, count: - 1 };
     /**
      * Used to check type of this or extended instance.
      * This value should not be changed by user.
      */
-    public isBufferAttribute = true;
+    isBufferAttribute = true;
     /**
      * When the data is cleared from memory, this mark will be set true.
      * @deprecated never updated, always `false`.
      */
-    public isDataDropped = false;
+    isDataDropped = false;
     /**
      * Remove the data after it is uploaded to GPU memory.
      * @deprecated will not take effect
      */
-    public removeDataAfterUpload = false;
+    removeDataAfterUpload = false;
     /**
      * A version number, incremented every time the {@link needsUpdate | needsUpdate } property is set to true.
      */
-    public version = 0;
+    version = 0;
 
     constructor(array: T, itemSize: number) {
         super();
@@ -102,17 +102,17 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * @ignore
      */
-    public getLayoutKey(): string {
+    getLayoutKey(): string {
         return this.array.constructor.name + this.itemSize;
     }
     /**
      * A callback function for updating. It will be called after {@link needsUpdate | needsUpdate }.
      */
-    public onUpdateCallback = function () { };
+    onUpdateCallback = function () { };
     /**
      * A callback function for uploading. It will be called after the data uploaded to WebGL.
      */
-    public onUploadCallback = function () { };
+    onUploadCallback = function () { };
     /**
      * When this attribute is set to true, the {@link version | version } will be change.
      */
@@ -125,10 +125,10 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Change the data of {@link array | array }.
      */
-    public setArray(array: T) {
+    setArray(array: T) {
         this.dataVersion++;
         this.count = array !== undefined ? array.length / this.itemSize : 0;
-        ContentBridge.bufferAttributeSetData(this, array, this.itemSize, this.count);
+        ContentBridge.bufferAttributeSetData(this, array, this.itemSize, this.count, this.normalized);
         if (!hasManagedContentAPI()) {
             this._array = array;
         } else {
@@ -142,7 +142,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Change the state of {@link dynamic | dynamic }.
      */
-    public setDynamic(value: boolean) {
+    setDynamic(value: boolean) {
         this.dynamic = value;
         return this;
     }
@@ -150,7 +150,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * Copy the data to this object from source.
      * @param { BufferAttribute } source the data source.
      */
-    public copy(source: BufferAttribute) {
+    copy(source: BufferAttribute) {
         this.array = source.array.slice() as T;
         this.itemSize = source.itemSize;
         this.count = source.count;
@@ -164,7 +164,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * @param { BufferAttribute } attribute the start position of this array.
      * @param { number } index2 the start position of source array.
      */
-    public copyAt(index1: number, attribute: BufferAttribute, index2: number) {
+    copyAt(index1: number, attribute: BufferAttribute, index2: number) {
         index1 *= this.itemSize;
         index2 *= attribute.itemSize;
         for (let i = 0, l = this.itemSize; i < l; i++) {
@@ -176,7 +176,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * Copy the data from given array to this {@link array | array }.
      * @param { TypedArray } array the start position of this array.
      */
-    public copyArray(array: TypedArray | number[]) {
+    copyArray(array: TypedArray | number[]) {
         this.array.set(array);
         return this;
     }
@@ -184,7 +184,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * Copy the data from given array to this {@link array | array }.
      * @param { Color[] } colors the start position of this array.
      */
-    public copyColorArray(colors: Color[]) {
+    copyColorArray(colors: Color[]) {
         const array = this.array;
         let offset = 0;
         for (let i = 0, l = colors.length; i < l; i++) {
@@ -199,7 +199,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * Copy the data which is type of Vector2 from given array to this {@link array | array }.
      * @param { Vector2[] } vectors the start position of this array.
      */
-    public copyVector2Array(vectors: Vector2[]) {
+    copyVector2Array(vectors: Vector2[]) {
         const array = this.array;
         let offset = 0;
         for (let i = 0, l = vectors.length; i < l; i++) {
@@ -213,7 +213,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * Copy the data which is type of Vector3 from given array to this {@link array | array }.
      * @param { Vector3[] } vectors the start position of this array.
      */
-    public copyVector3Array(vectors: Vector3[]) {
+    copyVector3Array(vectors: Vector3[]) {
         const array = this.array;
         let offset = 0;
         for (let i = 0, l = vectors.length; i < l; i++) {
@@ -228,7 +228,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * Copy the data which is type of Vector4 from given array to this {@link array | array }.
      * @param { Vector4[] } vectors the start position of this array.
      */
-    public copyVector4Array(vectors: Vector4[]) {
+    copyVector4Array(vectors: Vector4[]) {
         const array = this.array;
         let offset = 0;
         for (let i = 0, l = vectors.length; i < l; i++) {
@@ -245,7 +245,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
      * @param { ArrayLike<number> } value an Array or TypedArray from which to copy values.
      * @param { number } offset (optional) index of the {@link array | array } at which to start copying.
      */
-    public set(value: ArrayLike<number>, offset?: number) {
+    set(value: ArrayLike<number>, offset?: number) {
         if (offset === undefined) {
             offset = 0;
         }
@@ -255,59 +255,59 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Return the x component of all items.
      */
-    public getX(index: number) {
+    getX(index: number) {
         return this.array[index * this.itemSize];
     }
     /**
      * Set the x component of the vector at the given index.
      */
-    public setX(index: number, x: number) {
+    setX(index: number, x: number) {
         this.array[index * this.itemSize] = x;
         return this;
     }
     /**
      * Return the y component of all vectors.
      */
-    public getY(index: number) {
+    getY(index: number) {
         return this.array[index * this.itemSize + 1];
     }
     /**
      * Set the y component of the vector at the given index.
      */
-    public setY(index: number, y: number) {
+    setY(index: number, y: number) {
         this.array[index * this.itemSize + 1] = y;
         return this;
     }
     /**
      * Return the z component of all vectors.
      */
-    public getZ(index: number) {
+    getZ(index: number) {
         return this.array[index * this.itemSize + 2];
     }
     /**
      * Set the z component of the vector at the given index.
      */
-    public setZ(index: number, z: number) {
+    setZ(index: number, z: number) {
         this.array[index * this.itemSize + 2] = z;
         return this;
     }
     /**
      * Return the w component of all vectors.
      */
-    public getW(index: number) {
+    getW(index: number) {
         return this.array[index * this.itemSize + 3];
     }
     /**
      * Set the w component of the vector at the given index.
      */
-    public setW(index: number, w: number) {
+    setW(index: number, w: number) {
         this.array[index * this.itemSize + 3] = w;
         return this;
     }
     /**
      * Set the x and y component of the vector at the given index.
      */
-    public setXY(index: number, x: number, y: number) {
+    setXY(index: number, x: number, y: number) {
         const array = this.array;
         index *= this.itemSize;
         array[index + 0] = x;
@@ -317,7 +317,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Set the x , y and z component of the vector at the given index.
      */
-    public setXYZ(index: number, x: number, y: number, z: number) {
+    setXYZ(index: number, x: number, y: number, z: number) {
         const array = this.array;
         index *= this.itemSize;
         array[index + 0] = x;
@@ -328,7 +328,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Set all component of the vector at the given index.
      */
-    public setXYZW(index: number, x: number, y: number, z: number, w: number) {
+    setXYZW(index: number, x: number, y: number, z: number, w: number) {
         const array = this.array;
         index *= this.itemSize;
         array[index + 0] = x;
@@ -340,7 +340,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Return the vector2 at specific position.
      */
-    public getVector2(index: number, out = new Vector2()) {
+    getVector2(index: number, out = new Vector2()) {
         const array = this.array;
         const i = index * this.itemSize;
         out.x = array[i];
@@ -350,20 +350,20 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * @ignore
      */
-    public onUpload(callback: () => void) {
+    onUpload(callback: () => void) {
         this.onUploadCallback = callback;
         return this;
     }
     /**
      * Return a clone of this object.
      */
-    public clone(): BufferAttribute<T> {
+    clone(): BufferAttribute<T> {
         return new BufferAttribute(this.array, this.itemSize).copy(this);
     }
     /**
      * @ignore
      */
-    public serialize(ctx: Serializer) {
+    serialize(ctx: Serializer) {
         ctx.puts<BufferAttribute>(['itemSize', 'normalized', 'count', 'dynamic']);
         ctx.putRaw('array', ctx.collectBuffer(this.array));
         ctx.putRaw('updateRange', ctx.deepClone(this.updateRange));
@@ -371,7 +371,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * @ignore
      */
-    public deserialize(ctx: Deserializer) {
+    deserialize(ctx: Deserializer) {
         ctx.reads<BufferAttribute>(['itemSize', 'normalized', 'count', 'dynamic']);
         this.array = ctx.getBuffer(ctx.readRaw('array')) as T;
         this.updateRange = ctx.readRaw('updateRange');
@@ -379,7 +379,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
     /**
      * Release the memory which is occupied by this instance from CPU and GPU.
      */
-    public destroy() {
+    destroy() {
         super.destroy();
         ContentBridge.bufferAttributeDestroy(this);
         // we must set this empty, because the detached arraybuffer still reference data
@@ -387,7 +387,7 @@ export class BufferAttribute<T extends TypedArray = TypedArray> extends ElementE
         this._array = new Float32Array() as T;
     }
 
-    public freeGPU() {
+    freeGPU() {
         ContentBridge.bufferAttributeFreeGPU(this);
     }
 }